Removed start script to different file.
